requirements.txt
pip install -r requirements.txt
requests
web3
websocket-client
webstompy